days_off/requirements.txt
Django
django-admin
django-bootstrap-v5
paho-mqtt